+++ src/wetterdienst/provider/dmi/observation/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.dmi.observation.api import DmiObservationRequest
+from wetterdienst.provider.dmi.observation.metadata import DmiObservationMetadata
+
+__all__ = [
+    "DmiObservationMetadata",
+    "DmiObservationRequest",
+]